Move examples to separate folder
Closes #27
